Move the BlaChecks.py to subfolder #162
