+++ rpmlint/checks/AbstractCheck.py
