+++ rpmlint/checks/NamingPolicyCheck.py
@@ -9,7 +9,7 @@
 
 import re
 
-from rpmlint.AbstractCheck import AbstractCheck
+from rpmlint.checks.AbstractCheck import AbstractCheck
 
 
 class NamingPolicyNotAppliedException(Exception):
