+++ test/test_rpmfile.py
@@ -1,6 +1,6 @@
 import pytest
+from rpmlint.checks.RpmFileCheck import RpmFileCheck
 from rpmlint.Filter import Filter
-from rpmlint.RpmFileCheck import RpmFileCheck
 
 from Testing import CONFIG, get_tested_package
