+++ test/test_tags.py
@@ -1,6 +1,6 @@
 import pytest
+from rpmlint.checks.TagsCheck import TagsCheck
 from rpmlint.Filter import Filter
-from rpmlint.TagsCheck import TagsCheck
 
 from Testing import CONFIG, get_tested_package
